Store log files for test CI artefacts

+++ Beam/Classes/Helpers/Utils/Logger.swift
@@ -31,14 +31,32 @@ final class Logger {
     private let hideCategories: [LogCategory] = [.web, .coredataDebug, .documentDebug]
     private let hideLumberCategories: [LogCategory] = [.documentDebug]
 
-    private var ddFileLogger: DDFileLogger = DDFileLogger()
+    private var ddFileLogger: DDFileLogger?
+
     private func configure() {
-        DDLog.add(ddFileLogger)
+//        let documentsDirectory = NSSearchPathForDirectoriesInDomains(FileManager.SearchPathDirectory.applicationDirectory,
+//                                                                     FileManager.SearchPathDomainMask.userDomainMask,
+//                                                                     true).first
+//        let fileManager = DDLogFileManagerDefault(logsDirectory: documentsDirectory)
+
+        if let ddFileLogger = ddFileLogger {
+            DDLog.remove(ddFileLogger)
+        }
+
+//        ddFileLogger = DDFileLogger(logFileManager: fileManager)
+
+        ddFileLogger = DDFileLogger()
+
+        if let ddFileLogger = ddFileLogger {
+            DDLog.add(ddFileLogger)
+        }
     }
 
     func removeFiles() {
+        guard let ddFileLogger = ddFileLogger else { return }
+
         ddFileLogger.rollLogFile(withCompletion: {
-            for filename: String in self.ddFileLogger.logFileManager.sortedLogFilePaths {
+            for filename: String in ddFileLogger.logFileManager.sortedLogFilePaths {
                 do {
                     try FileManager.default.removeItem(atPath: filename)
                 } catch {
@@ -51,12 +69,20 @@ final class Logger {
 
     private init() {
         configure()
+
+        guard let ddFileLogger = ddFileLogger else { return }
+
+        // swiftlint:disable:next print
+        print("Storing log files to \(ddFileLogger.currentLogFileInfo?.filePath ?? "-")")
+
         ddFileLogger.logFileManager.maximumNumberOfLogFiles = 2
         ddFileLogger.maximumFileSize = 1024 * 64 // 64k
         ddFileLogger.rollingFrequency = 3600 * 24 * 7 // 1 week
     }
 
     var logFileData: Data {
+        guard let ddFileLogger = ddFileLogger else { return Data() }
+
         let logFilePaths = ddFileLogger.logFileManager.sortedLogFilePaths
         var logFileDataArray = Data()
         for logFilePath in logFilePaths.reversed() {
